@@ -2,7 +2,7 @@
 name = "divbuf"
 version = "0.4.1"
 authors = ["Alan Somers <asomers@gmail.com>"]
-edition = "2018"
+edition = "2021"
 license = "MIT"
 readme = "README.md"
 repository = "https://github.com/asomers/divbuf"
@@ -16,6 +16,7 @@ exclude       = [
     ".gitignore",
     ".travis.yml"
 ]
+rust-version = "1.80"
 
 [package.metadata.docs.rs]
 features = ["experimental"]
@@ -27,4 +28,3 @@ experimental = []
 [dependencies]
 
 [dev-dependencies]
-lazy_static = "1.1"
